Temp fix for perm errors

+++ runtime/commands/unlisted/owo.js
@@ -8,6 +8,11 @@ module.exports =
 	run: async (options) =>
 	{
 		BotActions.send(options, "OwO");
-		options.message.delete().catch();
+		
+			let myMember = options.guild.members.cache.get(VarClient.user.id);
+		let hasPerm = myMember.permissions.has("MANAGE_MESSAGES");
+		if (hasPerm)
+			options.message.delete().catch();
+		// todo: BotTools delete message function
 	}
 }
